python: Always rebuild generated files.
Currently, two generated python files (dirs.py and version.py) are
included in dist tarball and not rebuilt if you use the release tarball.

If dirs.py is not rebuilt, it causes problems when a non-default value for
prefixes are specified in ./configure.
For example, ipsec cannot start if /usr/local is not used as prefix directory,
since dirs.py includes /usr/local.

If version.py is not rebuilt, it causes problem when a non-default value for
version suffix is specified in ./configure, since the default one will be used
instead.

Fixes: 9e6d43ef3215 ("rhel: Make the version, displayed to the user, customizable.")
Reported-at: https://bugzilla.redhat.com/2055576
Signed-off-by: Timothy Redaelli <tredaelli@redhat.com>